requirements.txt
openai==1.30.1
azure-search-documents==11.6.0b4
azure-storage-blob==12.20.0
azure-identity==1.17.0
fastapi==0.111.0
uvicorn[standard]==0.30.1
streamlit==1.35.0
langchain==0.2.5
langchain-openai==0.1.9
pandas==2.2.2
tiktoken==0.7.0
python-dotenv==1.0.1
httpx==0.27.0
pydantic==2.7.4
slowapi==0.1.9
pytest==8.2.2
pytest-asyncio==0.23.7
httpx==0.27.0
